==> notebooks/02-baseline-and-sweep.ipynb <==

import lightgbm as lg
import pandas as pd
import lightgbm as lgb
from sklearn.metrics import mean_squared_error
from sklearn import metrics
from sklearn.metrics import average_precision_score
import numpy as np
import wandb
from wandb.lightgbm import wandb_callback, log_summary
from sklearn.metrics import accuracy_score, classification_report, roc_auc_score, f1_score, precision_recall_fscore_support, roc_curve
from sklearn.metrics import RocCurveDisplay, PrecisionRecallDisplay
import matplotlib.pyplot as plt


1. Load data

train_df = pd.read_csv('/Users/favea/Downloads/swiss-data/train_df.csv' ,index_col=0)
before_flights = pd.read_csv('/Users/favea/Downloads/swiss-data/previous_delayed_same_day.csv').drop('Unnamed: 0',axis=1)
train_df = train_df.merge(before_flights, left_on='id_flight', right_on='wh_fleg_leg_i')

train_df['departure_hour'] = train_df.scheduled_time_departure.apply(lambda x: pd.to_datetime(x).hour)
train_df['departure_minutes'] = train_df.scheduled_time_departure.apply(lambda x: pd.to_datetime(x).minute)

columns_to_remove = ['ac_subtype','wh_fleg_leg_i','id_flight','carrier','season_trip','delay_minutes','scheduled_time_departure','year']


train_df = train_df.drop(columns_to_remove,axis=1)
train_df = train_df.drop('ac_registration_code',axis=1)
x_train, y_train = train_df.drop('delayed',axis=1), train_df.loc[:,'delayed']
x_train.origin = x_train.origin.astype('category')
x_train.destination = x_train.destination.astype('category')
x_train.previous_is_delayed_same_day = x_train.previous_is_delayed_same_day.astype('category')



lgb_train = lgb.Dataset(x_train, y_train, categorical_feature = ['origin','destination','previous_is_delayed_same_day'], free_raw_data=False)

dev_df = pd.read_csv('/Users/favea/Downloads/swiss-data/dev_df.csv', index_col=0)
dev_df = dev_df.merge(before_flights, left_on='id_flight', right_on='wh_fleg_leg_i')

dev_df['departure_hour'] = dev_df.scheduled_time_departure.apply(lambda x: pd.to_datetime(x).hour)
dev_df['departure_minutes'] = dev_df.scheduled_time_departure.apply(lambda x: pd.to_datetime(x).minute)


dev_df = dev_df.drop(columns_to_remove,axis=1)
dev_df = dev_df.drop('ac_registration_code',axis=1)

x_dev, y_dev = dev_df.drop('delayed',axis=1), dev_df.loc[:,'delayed']
x_dev.origin = x_dev.origin.astype('category')
x_dev.destination = x_dev.destination.astype('category')

x_dev.previous_is_delayed_same_day = x_dev.previous_is_delayed_same_day.astype('category')
lgb_dev = lgb.Dataset(x_dev, y_dev, reference=lgb_train, categorical_feature =['origin','destination','previous_is_delayed_same_day'], free_raw_data=False)

2. Train baseline

# specify your configurations as a dict
params = {
    'boosting_type': 'gbdt',
    'objective': 'binary',
    'metric': ['auc','average_precision'],
    'feature_fraction':0.8, # ration of features to randomly select
    'bagging_fraction':0.8, # ratio of randomly selected sample of rows
    'bagging_freq':10, # frequency of iterations to update selected rows
    'num_leaves': 70,
    'max_depth': 8,
    'learning_rate': 0.02,
    'verbosity': 1,
    'scale_pos_weight': 20,
    'max_bin': 255,
}

wandb.init(project='swiss-delay-prediction', job_type= 'training-lightgbm-2-baseline', config=params, reinit=True)

gbm = lgb.train(params,
                lgb_train,
                num_boost_round=10_000,
                valid_sets=lgb_dev,
                valid_names=('validation'),
                callbacks=[wandb_callback()],
                early_stopping_rounds=20,
                )

log_summary(gbm, save_model_checkpoint=True) # save model checkpooint and metrics to WANDB

# predict
y_pred = gbm.predict(x_dev, num_iteration=gbm.best_iteration)

# save AUC 
roc_ac = roc_auc_score(y_dev, y_pred)
wandb.log({'val_roc_auc': roc_ac})

# Save confusion matrix
wandb.sklearn.plot_confusion_matrix(y_dev, (y_pred>=0.5).astype(int))

# Classification report on dev set 
d = classification_report(y_dev, (y_pred>=0.5).astype(int), output_dict=True)
sensitivity = d['1']['recall']
specificity = d['0']['recall']
f1_score = d['macro avg']['f1-score']

wandb.summary['sensitivity'] = sensitivity
wandb.summary['specificity'] = specificity
wandb.summary['f1_score'] = f1_score

# save PR-curve
display = PrecisionRecallDisplay.from_predictions(    
    y_dev,
    y_pred,
    color="darkorange",
    name = 'LightGBM'
)
plt.plot()
wandb.log({'PR-curve':wandb.Image(plt)})

# Save ROC curve
RocCurveDisplay.from_predictions(
    y_dev,
    y_pred,
    color="darkorange",
)
plt.plot([0, 1], [0, 1], "k--", label="chance level (AUC = 0.5)")
plt.axis("square")
plt.xlabel("False Positive Rate")
plt.ylabel("True Positive Rate")
plt.legend()

wandb.log({'ROC-curve':wandb.Image(plt)})

wandb.finish() # complete

## SWEEP: Hyperparameter tuning

### Suggestions for LightGBM

For heavily unbalanced datasets such as 1:10000:

max_bin: keep it only for memory pressure, not to tune (otherwise overfitting)

learning rate: keep it only for training speed, not to tune (otherwise overfitting)

n_estimators: must be infinite (like 9999999) and use early stopping to auto-tune (otherwise overfitting)

num_leaves: [7, 4095]

max_depth: [2, 63] and infinite (I personally saw metric performance increases with such 63 depth with small number of leaves on sparse unbalanced datasets)

scale_pos_weight: [1, 10000] (if over 10000, something might be wrong because I never saw it that good after 5000)

min_child_weight: [0.01, (sample size / 1000)] if you are using logloss (think about the hessian possible value rang
e before putting "sample size / 1000", it is dataset-dependent and loss-dependent)


subsample: [0.4, 1]


bagging_freq: only 1, keep as is (otherwise overfitting)

colsample_bytree: [0.4, 1]

is_unbalance: false (make your own weighting with scale_pos_weight)

USE A CUSTOM METRIC (to reflect reality without weighting, otherwise you have weights inside your metric with premade metrics like xgboost)

Never tune these parameters unless you have an explicit requirement to tune them:

Learning rate (lower means longer to train but more accurate, higher means smaller to train but less accurate)

Number of boosting iterations (automatically tuned with early stopping and learning rate)

Maximum number of bins (RAM dependent)

sweep_config = {
  "method" : "bayes",
  "metric": {
      "name": "avg_precision_dev",
      "goal": "maximize"
  },
  "parameters" : {
    "gamma" :{
      "min": 0.001,
      "max": 1.0
    },
    "min_child_weight" :{
      "min": 1,
      "max": 150
    },
    "scale_pos_weight":{
      "min": 1,
      "max": 30,
    },
    "num_leaves" :{
      "min": 10,
      "max": 100
    },
    "max_depth" :{
      "min": 5,
      "max": 40
    },
    "feature_fraction" :{
      "min": 0.3,
      "max": 1.0
    },
    "bagging_fraction" :{
      "min": 0.3,
      "max": 1.0
    },

    "lambda_l1":{"values":[0.5, 0.8, 1]},
    "lambda_l2":{"values":[0.5, 0.8, 1]},
    

  }
}


def train():     
    with wandb.init() as run:

        params = {
            'boosting_type': 'gbdt',
            'metric': ['auc','average_precision'],
            'objective': 'binary',
            'num_leaves': run.config['num_leaves'],
            'learning_rate': 0.008,
            'bagging_freq': 10,
            'verbosity': 1,
            'scale_pos_weight': run.config['scale_pos_weight'],
            'bagging_fraction': run.config['bagging_fraction'],
            'max_depth': run.config['max_depth'],
            'scale_pos_weight': run.config['scale_pos_weight'],
            'feature_fraction': run.config['feature_fraction'],
            'lambda_l2':run.config['lambda_l2'],
            'lambda_l1':run.config['lambda_l1'],
            'max_bin':255,

        }
    

        # Initialize and train LightGBM model
        gbm = lgb.train(params,
                        lgb_train,
                        num_boost_round=10000, #infinite
                        valid_sets=lgb_dev,
                        valid_names=('validation'),
                        callbacks=[wandb_callback()],
                        early_stopping_rounds=20,
                        )

        log_summary(gbm, save_model_checkpoint=True)
        


        # Log booster metrics
        run.summary["best_score"] = gbm.best_score
        run.summary["best_iteration"] = gbm.best_iteration
        
        # Get train and validation predictions
        y_dev_pred = gbm.predict(x_dev, num_iteration=gbm.best_iteration)
        y_train_pred = gbm.predict(x_train, num_iteration=gbm.best_iteration)


        # Log additional Train metrics
        false_positive_rate, true_positive_rate, thresholds = metrics.roc_curve(y_train, y_train_pred) 

        avg_precision_train = average_precision_score(y_train, y_train_pred)
        run.summary['train_avg_precision'] = avg_precision_train
        run.summary['train_auc'] = metrics.auc(false_positive_rate, true_positive_rate)
        run.summary['train_log_loss'] = -(y_train * np.log(y_train_pred) + (1-y_train) * np.log(1-y_train_pred)).sum() / len(y_train)

        # Log additional Validation metrics
        avg_precision_dev = average_precision_score(y_dev, y_dev_pred)
        run.summary['avg_precision_dev'] = avg_precision_dev
        run.summary["val_auc"] = metrics.roc_auc_score(y_dev, y_dev_pred)
        run.summary["val_acc_0.5"] = metrics.accuracy_score(y_dev, np.where(y_dev_pred >= 0.5, 1, 0))
        run.summary["val_log_loss"] = -(y_dev * np.log(y_dev_pred) 
                                             + (1-y_dev) * np.log(1-y_dev_pred)).sum() / len(y_dev)

        d = classification_report(y_dev, np.where(y_dev_pred >= 0.5, 1, 0), output_dict=True)
        sensitivity = d['1']['recall']
        specificity = d['0']['recall']
        f1_score = d['macro avg']['f1-score']

        run.summary['val_sensitivity_0.5'] = sensitivity
        run.summary['val_specificity_0.5'] = specificity
        run.summary['val_f1_score'] = f1_score


        # Log plots
        RocCurveDisplay.from_predictions(
            y_dev,
            y_dev_pred,
            color="darkorange",
        )
        plt.plot([0, 1], [0, 1], "k--", label="chance level (AUC = 0.5)")
        plt.axis("square")
        plt.xlabel("False Positive Rate")
        plt.ylabel("True Positive Rate")
        plt.legend()
        run.log({'ROC-curve-dev':wandb.Image(plt)})

        display = PrecisionRecallDisplay.from_predictions(    
            y_dev,
            y_dev_pred,
            color="darkorange",
            name = 'LightGBM'
        )
        plt.plot()
        plt.xlabel("Precision")
        plt.ylabel("Recall")
        plt.legend()
        run.log({'PR-curve-dev':wandb.Image(plt)})
        try:
            wandb.sklearn.plot_confusion_matrix(y_dev, (y_dev_pred>=0.5).astype(int))
        except:
            pass


sweep_id = wandb.sweep(sweep_config, project="swiss-delay-prediction")

count = 1000 # number of runs to execute
wandb.agent(sweep_id, function=train, count=count)
