==> notebooks/04-smote-imbalance.ipynb <==
## Using SMOTE-NC + Random Undersampling
There's a Python Script for simplicity

import pandas as pd
import numpy as np
import imblearn
from collections import Counter
from imblearn.over_sampling import SMOTENC
from imblearn.under_sampling import RandomUnderSampler
from imblearn.pipeline import Pipeline
from imblearn.combine import SMOTETomek

train_df = pd.read_csv('/Users/favea/Downloads/swiss-data/train_df.csv', index_col=0).drop('Unnamed: 0',axis=1).iloc[:1000]

X,y = train_df.drop('delayed',axis=1), train_df.loc[:,'delayed']
Counter(y) # Distribution of delayed in Y before sampling

# define pipeline
over = SMOTENC(categorical_features=[0, 1, 2,4,9,], random_state=12345, 
                sampling_strategy=0.5,
                n_jobs=-1)
under = RandomUnderSampler(random_state=12345, sampling_strategy=0.5)
steps = [('o', over), ('u', under)]
pipeline = Pipeline(steps=steps)
# transform the dataset
X, y = pipeline.fit_resample(X, y)
Counter(y) #  Distribution of delayed in Y after sampling
