==> notebooks/05-merged-weather-data.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from lightgbm import LGBMClassifier
import lightgbm as lgb
import pandas as pd
import os
from sklearn.metrics import classification_report

data_path = '/Users/favea/Downloads/swiss-data'
weather_files = [os.path.join(data_path,f) for f in os.listdir(data_path) if 'weather' in f] # get all files about weather

cleaned_df = pd.read_csv(os.path.join(data_path,'12122022_cleaned.csv'),index_col=0) # get main df 

# Append all weather df together
df_weather = None
for wf in weather_files:
    print(wf)
    if df_weather is None:
        df_weather = pd.read_csv(wf)
        print(df_weather.shape)
    else:
        d = pd.read_csv(wf)
        print(d.shape)
        print('Appending')
        df_weather = df_weather.append(d)



merged = pd.merge(left = cleaned_df, right = df_weather, on = ['origin','destination','scheduled_time_departure'])

merged = merged.dropna().drop(['year','delay_minutes','scheduled_time_departure','Unnamed: 0'],axis=1)

Train model without weather features

weather_features = ['temperature','precipitation', 'windSpeed', 'dewpoint', 'humidity']

X,y = merged.drop('delayed',1), merged.loc[:,'delayed']

X = X.drop(weather_features+['id_flight','carrier','ac_subtype','season_trip'],axis=1)
X.origin = X.origin.astype('category')
X.destination = X.destination.astype('category')
X.ac_registration_code = X.ac_registration_code.astype('category') 

seed = 7
test_size = 0.20
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed, stratify=y)
model =LGBMClassifier(learning_rate=0.2,n_estimators=10_000, early_stopping_rounds=30)
gbm = model.fit(X_train, y_train, eval_set=(X_test,y_test), eval_metric ='average_precision') 

lgb.plot_importance(gbm)

lgb.plot_metric(gbm,'average_precision')

y_test_pred = gbm.predict(X_test, num_iteration = gbm.best_iteration_)
print(classification_report(y_test, (y_test_pred>=0.5).astype(int)))

Train model with weather features

X,y = merged.drop('delayed',1), merged.loc[:,'delayed']
X = X.drop(['id_flight','carrier','ac_subtype','season_trip'],axis=1)

X.origin = X.origin.astype('category')
X.destination = X.destination.astype('category')
X.ac_registration_code = X.ac_registration_code.astype('category')

seed = 7
test_size = 0.2
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_size, random_state=seed, stratify=y)

model = LGBMClassifier(learning_rate=0.2,n_estimators=10_000, early_stopping_rounds=30)
gbm = model.fit(X_train, y_train, eval_set=(X_test,y_test), eval_metric ='average_precision')


lgb.plot_importance(gbm)

lgb.plot_metric(gbm,'average_precision')

y_test_pred = gbm.predict(X_test, num_iteration = gbm.best_iteration_)
print(classification_report(y_test, (y_test_pred>=0.5).astype(int)))
